move types to another file
