+++ features/aave/manage/components/index.ts
@@ -1,2 +1,3 @@
 export * from '../../../earn/aave/components/ManageSectionComponent'
 export * from './ManageSectionModal'
+export { ManageSectionComponentProps } from '../../common/StrategyConfigTypes'
